Update rust-toolchain
Upgrade Rust version to 1.74.0 in rust-toolchain file

This commit updates the Rust toolchain version from 1.70.0 to 1.74.0 in the rust-toolchain file. The upgrade is necessary as the Carpe project's minimum dependency version has been identified as 1.74.0. This change ensures compatibility with all required dependencies and leverages the latest Rust features and improvements for better performance and stability.